Extract `EventHandler` from NativeProxy.h
