+++ packages/react-native-reanimated/android/src/main/cpp/reanimated/android/NativeProxy.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <reanimated/NativeModules/ReanimatedModuleProxy.h>
-#include <reanimated/Tools/ReanimatedSystraceSection.h>
 
 #include <worklets/android/WorkletsModule.h>
 
@@ -44,39 +43,6 @@ class AnimationFrameCallback : public HybridClass<AnimationFrameCallback> {
       : callback_(std::move(callback)) {}
 
   std::function<void(double)> callback_;
-};
-
-class EventHandler : public HybridClass<EventHandler> {
- public:
-  static auto constexpr kJavaDescriptor =
-      "Lcom/swmansion/reanimated/nativeProxy/EventHandler;";
-
-  void receiveEvent(
-      jni::alias_ref<JString> eventKey,
-      jint emitterReactTag,
-      jni::alias_ref<react::WritableMap> event) {
-    ReanimatedSystraceSection s("EventHandler::receiveEvent");
-    handler_(eventKey, emitterReactTag, event);
-  }
-
-  static void registerNatives() {
-    javaClassStatic()->registerNatives({
-        makeNativeMethod("receiveEvent", EventHandler::receiveEvent),
-    });
-  }
-
- private:
-  friend HybridBase;
-
-  explicit EventHandler(std::function<void(
-                            jni::alias_ref<JString>,
-                            jint emitterReactTag,
-                            jni::alias_ref<react::WritableMap>)> handler)
-      : handler_(std::move(handler)) {}
-
-  std::function<
-      void(jni::alias_ref<JString>, jint, jni::alias_ref<react::WritableMap>)>
-      handler_;
 };
 
 class SensorSetter : public HybridClass<SensorSetter> {
